Fill out NPC sheet sections
Add most of the sheet sections in editable and normal mode.
