+++ module/bwactor.ts
@@ -1,4 +1,4 @@
-import { ShadeString, updateTestsNeeded, StringIndexedObject, getWorstShadeString } from "./helpers.js";
+import { ShadeString, StringIndexedObject } from "./helpers.js";
 import { ArmorRootData, DisplayClass, ItemType, Trait, TraitDataRoot, ReputationDataRoot, PossessionRootData, BWItemData } from "./items/item.js";
 import { SkillDataRoot } from "./items/skill.js";
 import * as constants from "./constants.js";
@@ -39,8 +39,13 @@ export class BWActor extends Actor {
     prepareData(): void {
         super.prepareData();
         if (this.data.type === "character") {
-            this._prepareCharacterData();
+            BWCharacter.prototype.bindCharacterFunctions.call(this);
+        } else {
+            Npc.prototype.bindNpcFunctions.call(this);
         }
+
+        this._prepareActorData();
+        this.prepareTypeSpecificData();
     }
 
     getForkOptions(skillName: string): { name: string, amount: number }[] {
@@ -65,36 +70,6 @@ export class BWActor extends Actor {
             });
     }
 
-    async updatePtgs(): Promise<this> {
-        const accessorBase = "data.ptgs.wound";
-        const forte = parseInt(this.data.data.forte.exp, 10) || 1;
-        const mw = this.data.data.mortalWound || 15;
-        const su = Math.floor(forte / 2) + 1;
-
-        const wounds = BWActor._calculateThresholds(mw, su, forte);
-        const updateData = {};
-        let woundType = "bruise";
-        for (let i = 1; i <= 16; i ++ ) {
-            if (i < wounds.su) {
-                woundType = "bruise";
-            } else if (i < wounds.li) {
-                woundType = "superficial";
-            } else if (i < wounds.mi) {
-                woundType = "light";
-            } else if (i < wounds.se) {
-                woundType = "midi";
-            } else if (i < wounds.tr) {
-                woundType = "severe";
-            } else if (i < wounds.mo) {
-                woundType = "traumatic";
-            } else {
-                woundType = "mortal";
-            }
-            updateData[`${accessorBase}${i}.threshold`] = woundType;
-        }
-        return this.update(updateData);
-    }
-
     private _addRollModifier(rollName: string, modifier: RollModifier, onlyNonZero = false) {
         rollName = rollName.toLowerCase();
         if (onlyNonZero && !modifier.dice && !modifier.obstacle) {
@@ -111,81 +86,13 @@ export class BWActor extends Actor {
         return (this.data.rollModifiers[rollName.toLowerCase()] || []).concat(this.data.rollModifiers.all || []);
     }
 
-    private static _calculateThresholds(mo: number, su: number, forte: number):
-        { su: number, li: number, mi: number, se: number, tr: number, mo: number } {
-        const maxGap = Math.ceil(forte / 2.0);
-        const tr = Math.min(mo - 1, su + (maxGap * 4));
-        const se = Math.min(tr - 1, su + (maxGap * 3));
-        const mi = Math.min(se - 1, su + (maxGap * 2));
-        const li = Math.min(mi - 1, su + (maxGap));
+    prepareTypeSpecificData(): void { return; }
 
-        return { su, li, mi, se, tr, mo };
-    }
-
-    private _prepareCharacterData() {
-        BWCharacter.prototype.bindCharacterFunctions.call(this);
-        if (!this.data.data.settings) {
-            this.data.data.settings = {
-                onlySuccessesCount: 'Faith, Resources, Perception',
-                showSettings: false,
-                roundUpHealth: false,
-                roundUpMortalWound: false,
-                roundUpReflexes: false,
-                armorTrained: false,
-                ignoreSuperficialWounds: false,
-                showBurner: false
-            };
-        }
+    private _prepareActorData() {
         this.data.rollModifiers = {};
         this.data.callOns = {};
-        this._calculatePtgs();
+        
         this._calculateClumsyWeight();
-        const woundDice = this.data.data.ptgs.woundDice || 0;
-        updateTestsNeeded(this.data.data.will, false, woundDice);
-        updateTestsNeeded(this.data.data.power, false, woundDice);
-        updateTestsNeeded(this.data.data.perception, false, woundDice);
-        updateTestsNeeded(this.data.data.agility, false, woundDice);
-        updateTestsNeeded(this.data.data.forte, false, woundDice);
-        updateTestsNeeded(this.data.data.speed, false, woundDice);
-        updateTestsNeeded(this.data.data.health);
-        updateTestsNeeded(this.data.data.steel, true, woundDice);
-        updateTestsNeeded(this.data.data.circles);
-        updateTestsNeeded(this.data.data.resources, true, -1);
-        updateTestsNeeded(this.data.data.custom1);
-        updateTestsNeeded(this.data.data.custom2);
-
-        this.data.data.maxSustained = parseInt(this.data.data.will.exp) - (this.data.data.ptgs.woundDice || 0) - 1;
-        this.data.data.maxObSustained = parseInt(this.data.data.forte.exp) - (this.data.data.ptgs.woundDice || 0) - this.data.data.forteTax - 1;
-
-        const unRoundedReflexes =
-            (parseInt(this.data.data.perception.exp, 10) +
-            parseInt(this.data.data.agility.exp, 10) +
-            parseInt(this.data.data.speed.exp, 10)) / 3.0;
-        this.data.data.reflexesExp = (this.data.data.settings.roundUpReflexes ?
-            Math.ceil(unRoundedReflexes) : Math.floor(unRoundedReflexes))
-            - (this.data.data.ptgs.woundDice || 0);
-        const shades = [this.data.data.perception.shade, this.data.data.agility.shade, this.data.data.speed.shade];
-        this.data.data.reflexesShade = getWorstShadeString(getWorstShadeString(shades[0], shades[1]), shades[2]);
-        if (this.data.data.reflexesShade === "B") {
-            this.data.data.reflexesExp += shades.filter(s => s !== "B").length;
-        } else if (this.data.data.reflexesShade === "G") {
-            this.data.data.reflexesExp += shades.filter(s => s === "W").length;
-        }
-
-        const unRoundedMortalWound =
-            (parseInt(this.data.data.power.exp, 10) + parseInt(this.data.data.forte.exp, 10)) / 2 + 6;
-        this.data.data.mortalWound = this.data.data.settings.roundUpMortalWound ?
-            Math.ceil(unRoundedMortalWound) : Math.floor(unRoundedMortalWound);
-        if (this.data.data.power.shade !== this.data.data.forte.shade) {
-            this.data.data.mortalWound += 1;
-        }
-        this.data.data.mortalWoundShade = getWorstShadeString(this.data.data.power.shade, this.data.data.forte.shade);
-
-        this.data.data.hesitation = 10 - parseInt(this.data.data.will.exp, 10);
-
-        this.data.successOnlyRolls = (this.data.data.settings.onlySuccessesCount || '')
-            .split(',')
-            .map(s => s.trim().toLowerCase());
 
         this.data.forks = [];
         this.data.wildForks = [];
@@ -267,36 +174,6 @@ export class BWActor extends Actor {
         return this.data.callOns[roll.toLowerCase()] || [];
     }
 
-    private _calculatePtgs() {
-        let suCount = 0;
-        let woundDice = 0;
-        this.data.data.ptgs.obPenalty = 0;
-        Object.entries(this.data.data.ptgs).forEach(([_key, value]) => {
-            const w = value as Wound;
-            const a = w.amount && parseInt(w.amount[0], 10);
-            if ((w && a)) {
-                switch (w.threshold) {
-                    case "superficial": suCount += a; break;
-                    case "light": woundDice += a; break;
-                    case "midi": woundDice += a*2; break;
-                    case "severe": woundDice += a*3; break;
-                    case "traumatic": woundDice += a*4; break;
-                }
-            }
-        });
-
-        if (suCount >= 3) {
-            woundDice ++;
-        } else if (!this.data.data.ptgs.shrugging && suCount >= 1 && !this.data.data.settings.ignoreSuperficialWounds) {
-            this.data.data.ptgs.obPenalty = 1;
-        }
-        if (this.data.data.ptgs.gritting && woundDice) {
-            woundDice --;
-        }
-
-        this.data.data.ptgs.woundDice = woundDice;
-    }
-
     // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any
     _onCreate(data: any, options: any, userId: string, context: any): void {
         super._onCreate(data, options, userId, context);
@@ -313,8 +190,9 @@ export class BWActor extends Actor {
         this.createOwnedItem(constants.bareFistData);
 
         if (this.data.type === "character") {
+            const character = this as unknown as BWCharacter & BWActor;
             setTimeout(() => {
-                if (this.data.data.settings.showBurner) {
+                if (character.data.data.settings.showBurner) {
                     new Dialog({
                         title: "Launch Burner?",
                         content: "This is a new character. Would you like to launch the character burner?",
@@ -322,7 +200,8 @@ export class BWActor extends Actor {
                             yes: {
                                 label: "Yes",
                                 callback: () => {
-                                    CharacterBurnerDialog.Open(this);
+                                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                                    CharacterBurnerDialog.Open(this as any);
                                 }
                             },
                             later: {
@@ -356,6 +235,8 @@ export class BWActor extends Actor {
             untrainedAll: 0
         };
 
+        const asCharacter = this.data.type === "character" ? this as unknown as BWActor & BWCharacter : undefined;
+
         this.data.items.filter(i => i.type === "armor" && (i as unknown as ArmorRootData).data.equipped)
             .forEach(i => {
             const a = i as unknown as ArmorRootData;
@@ -384,7 +265,7 @@ export class BWActor extends Actor {
             }
 
 
-            if (!this.data.data.settings.armorTrained &&
+            if (asCharacter && !asCharacter.data.data.settings.armorTrained &&
                 (a.data.hasHelm || a.data.hasLeftArm || a.data.hasRightArm || a.data.hasTorso || a.data.hasLeftLeg || a.data.hasRightLeg)) {
                 // if this is more than just a shield
                 if (a.data.untrainedPenalty === "plate") {
@@ -399,7 +280,7 @@ export class BWActor extends Actor {
             }
         });
 
-        this.data.data.clumsyWeight = clumsyWeight;
+        if (asCharacter) { asCharacter.data.data.clumsyWeight = clumsyWeight; }
         const baseModifier = { optional: true, label: "Clumsy Weight" };
         this._addRollModifier("climbing", { obstacle: clumsyWeight.climbingPenalty, ...baseModifier }, true);
         this._addRollModifier("perception", { obstacle: clumsyWeight.helmetObPenalty,  ...baseModifier }, true);
@@ -462,7 +343,7 @@ export interface Common {
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export interface BWActorDataRoot extends ActorData<any> {
+export interface BWActorDataRoot extends ActorData<BWCharacterData | NpcData> {
     toolkits: PossessionRootData[];
     martialSkills: SkillDataRoot[];
     sorcerousSkills: SkillDataRoot[];
@@ -515,29 +396,6 @@ export interface DisplayProps {
     collapseSpells: boolean;
 }
 
-export interface Ptgs {
-    ptgs: {
-        wound1: Wound, wound2: Wound, wound3: Wound, wound4: Wound,
-        wound5: Wound, wound6: Wound, wound7: Wound, wound8: Wound,
-        wound9: Wound, wound10: Wound, wound11: Wound, wound12: Wound,
-        wound13: Wound, wound14: Wound, wound15: Wound, wound16: Wound,
-        woundNotes1: string,
-        woundNotes2: string,
-        woundNotes3: string,
-        shrugging: boolean, // shrug it off is active
-        gritting: boolean, // grit your teeth is active
-
-        // not persisted
-        obPenalty?: number,
-        woundDice?: number
-    };
-}
-
-interface Wound {
-    amount: string[]; // quirk relating to how radio button data is stored
-    threshold: string;
-}
-
 export interface ClumsyWeightData {
     agilityPenalty: number;
     speedObPenalty: number;
@@ -566,5 +424,6 @@ export interface NewItemData extends StringIndexedObject<any> {
 }
 
 // ====================== Character Class ===============================
-import { BWCharacter } from "./character.js";
+import { BWCharacter, BWCharacterData } from "./character.js";
+import { Npc, NpcData } from "./npc.js";
 export * from "./character.js";
