+++ templates/parts/npc-editable.hbs
@@ -1,0 +1,4 @@
+<div class="npc-editable" data-id="{{item._id}}" data-action="edit">
+    {{item.name}}
+    <i class="fas fa-trash" data-id="{{item._id}}" data-action="delete"></i>
+</div>
